+++ noxfile.py
@@ -39,7 +39,6 @@ def lint(session):
         "cerebras",
         "cohere",
         "dspy",
-        "dspy_2",
         "google_ai_studio",
         "groq",
         "instructor",
