+++ pyproject.toml
@@ -52,8 +52,7 @@ docs = ["playwright", "lazydocs", "nbformat", "nbconvert"]
 anthropic = ["anthropic>=0.18.0"]
 cerebras = ["cerebras-cloud-sdk"]
 cohere = ["cohere>=5.9.1,<5.9.3"]
-dspy = ["dspy>=0.1.5", "litellm<=1.49.1"]
-dspy_2 = ["dspy>=2.0.0"]
+dspy = ["dspy>=2.0.0"]
 google_ai_studio = ["google-generativeai>=0.8.3"]
 groq = ["groq>=0.13.0"]
 huggingface = ["huggingface-hub>=0.28.1"]
