+++ weave/trace/autopatch.py
@@ -61,7 +61,6 @@ def autopatch(settings: Optional[AutopatchSettings] = None) -> None:
     from weave.integrations.anthropic.anthropic_sdk import get_anthropic_patcher
     from weave.integrations.cerebras.cerebras_sdk import get_cerebras_patcher
     from weave.integrations.cohere.cohere_sdk import get_cohere_patcher
-    from weave.integrations.dspy.dspy_2_sdk import get_dspy_2_patcher
     from weave.integrations.dspy.dspy_sdk import get_dspy_patcher
     from weave.integrations.google_ai_studio.google_ai_studio_sdk import (
         get_google_genai_patcher,
@@ -89,7 +88,6 @@ def autopatch(settings: Optional[AutopatchSettings] = None) -> None:
     get_groq_patcher(settings.groq).attempt_patch()
     get_instructor_patcher(settings.instructor).attempt_patch()
     get_dspy_patcher(settings.dspy).attempt_patch()
-    get_dspy_2_patcher(settings.dspy).attempt_patch()
     get_cerebras_patcher(settings.cerebras).attempt_patch()
     get_cohere_patcher(settings.cohere).attempt_patch()
     get_google_genai_patcher(settings.google_ai_studio).attempt_patch()
@@ -106,7 +104,6 @@ def reset_autopatch() -> None:
     from weave.integrations.anthropic.anthropic_sdk import get_anthropic_patcher
     from weave.integrations.cerebras.cerebras_sdk import get_cerebras_patcher
     from weave.integrations.cohere.cohere_sdk import get_cohere_patcher
-    from weave.integrations.dspy.dspy_2_sdk import get_dspy_2_patcher
     from weave.integrations.dspy.dspy_sdk import get_dspy_patcher
     from weave.integrations.google_ai_studio.google_ai_studio_sdk import (
         get_google_genai_patcher,
@@ -134,7 +131,6 @@ def reset_autopatch() -> None:
     get_groq_patcher().undo_patch()
     get_instructor_patcher().undo_patch()
     get_dspy_patcher().undo_patch()
-    get_dspy_2_patcher().undo_patch()
     get_cerebras_patcher().undo_patch()
     get_cohere_patcher().undo_patch()
     get_google_genai_patcher().undo_patch()
